requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
astunparse==1.6.3
av==17.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
numpy==2.4.3
onnxruntime==1.24.4
openwakeword==0.6.0
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
protobuf==7.34.1
pycparser==3.0
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
wrapt==2.1.2
